color: extend linear transfer function
This change makes the scRGB color space transfer characteristics
encodable as "ext_linear". The key points are supporting negative and
greater than one values, and the implied luminances.

We avoid the need to add another TF just for scRGB. I am not aware of
any use case that would be broken by these differences to the H.273 code
point; material using H.273-linear code point should presumably work
just fine with the protocol "ext_linear". Still, the name "linear" is
left unused in case we need it later.

The luminances are chosen to represent typical SDR content. The minimum
luminance of zero anchors the black point, making it indifferent to the
choice of relative vs. absolute luminance.

When this TF is used in the absolute luminance context, it implies the
Windows-scRGB (HDR) convention of [1.0, 1.0, 1.0] = 80 cd/m².

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>