color: make invalid id illegal
This started from
https://gitlab.freedesktop.org/swick/wayland-protocols/-/merge_requests/90
where it was desireable for a compositor to send invalid image
description ids because it does not track and de-duplicate image
description records.

Concensus decided the opposite: the id number must be valid. The ability
to skip get_information request was seen valuable, and the effort to
implement id allocation was deemed not too high. It is also to the
majority taste to deliver image descriptions with valid id.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>