tests: Make build-only tests actually build-only
The goal behind the pedantic compiler tests are to ensure that the code
that wayland-scanner is generating can be compiled in pedantic mode by
the system C compiler.

Trying to execute the built tests may fail because of undefined symbols.
This affects certain platforms more than others; Linux/musl and Darwin
are examples of platforms that cannot execute binaries with undefined
symbols.  This meant tests needlessly failed on these platforms.

Signed-off-by: A. Wilcox <AWilcox@Wilcox-Tech.com>
Closes: #48, #228