color: st2084_pq dictates min/max_lum
ST 2084 a.k.a perceptual quantizer is an absolute luminance encoding
dealing in cd/m². This sets it apart for most other transfer
characteristics that are relative to the target display dynamic range.

Therefore the primary color volume minimum and maximum are fixed by the
ST 2084 specification, and cannot be changed. The reference white is not
fixed and remains settable.

ST 2084 allows displays to deviate from the specified EOTF due to
hardware limitations. How that works is not defined. Therefore code
point zero may produce more than 0 cd/m², but accoding to BT.2100 should
be less or equal to 0.005 cd/m². The difference between decoding with
the exact EOTF and then tone-mapping to display capability, and decoding
straight to display capability is mostly philosophical. The ideal
presentation is defined to be exactly as the ST 2084 specified EOTF.

I believe the image description should reflect this ideal presentation.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>