ext-zones: Add request to retrieve extents of server-side window frames
Positioning and aligning adjacent toplevel windows with server-side
decorations
would result in windows overlapping due to the compositor applying a
frame around
the window, which needs to be accounted for by clients when positioning
a window.

This adds a request to retrieve the extents of the server-side frame, so
that they
may be taken into consideration when positioning server decorated
windows relative
to one another.

Signed-off-by: Frank Praznik <frank.praznik@gmail.com>