color: MDCV min L < maxFALL <= maxCLL <= max L
These are the logical requirements. Also if the computation of maxFALL
and maxCLL follow the ANSI/CTA-861-I Annex P algorithms, then
necessarily maxFALL <= maxCLL.

On one hand, making these a protocol requirement causes clients to have
to check these themselves and fix things up in any arbitrary way, while
the compositors will do the exact same checks to send protocol errors.

On the other hand, clients checking these explicitly gives clients the
opportunity to employ a desired fix-up scheme and/or show a UI complaint
that the video content metadata is inconsistent. That would be more
difficult if only some compositors sent a graceful failure when these
requirements are not met.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>