color: update some protocol errors
We've opted to graceful failure the image description creation when bad
custom primaries or mastering display primaries are set. But we forgot
to update the errors related to that.

This renames 'invalid_primaries' to 'invalid_named_primaries',
clarifying that it is only possible to have such errors when clients set
primaries through the enums.

Also, this drops the 'invalid_mastering' error, as it is now unused in
the spec.

Signed-off-by: Leandro Ribeiro <leandro.ribeiro@collabora.com>