xdg-shell: Clarify xdg_surface.ack_configure behavior
The current wording is open to interpretation. The updated version
clarifies that the client is not required to commit the surface
immediately after receiving a configure event (for example it can
wait N milliseconds, paint and commit the surface) but it must still do
it within a reasonable timeframe.

The main motivation behind this behavior is that some Wayland
compositors use configure events to synchronize the server side double
buffered state with the client side state. If the client decides that
the received configure event can be postponed indefinetely, this can
leave the compositor with outdated current state, e.g. wrong position,
and so on.

Signed-off-by: Vlad Zahorodnii <vlad.zahorodnii@kde.org>