xdg-shell: relax grabbing popup parent requirements
Prior to b06650146e565ffe980a54ecd95aa6f7227e30d4 (xdg-shell: Delete
duplicate paragraph in xdg_popup) there were two paragraphs specifying
the requirements to the parent of a grabbing popup, which were
contradicting each other (see
https://gitlab.freedesktop.org/wayland/wayland-protocols/-/issues/82).
The referenced commit removed the paragraph which effectively stated
that a non-grabbing popup can be a grabbing popup's parent if there are
no explicit grabs already taken. However, as suggested by
https://gitlab.freedesktop.org/wayland/wayland-protocols/-/issues/82#note_1274459
and https://gitlab.freedesktop.org/wayland/weston/-/merge_requests/1607,
the intention was to specifically allow the mentioned case.

Aside from that, the current description suggests that a parent of a
grabbing popup must have the xdg_toplevel or the xdg_popup role. This
restriction means that a surface with some other role which can be a
parent of an xdg_popup can't be a parent of grabbing popup, which
doesn't seem to be intentional.

With this commit, the requirement for a grabbing popup's parent is that
it must be a grabbing popup if there is an existing explicit grab. A
sentence in the second-to-last paragraph which duplicated the current
description has been removed.

Additionally, this commit fixes a typo in the last paragraph (top most
-> topmost).

Signed-off-by: Kirill Primak <vyivel@eclair.cafe>